move decision datastructures and parsing
